+++ tests/scriptlets/set-session-storage-item.test.js
@@ -163,6 +163,20 @@ if (isSafariBrowser()) {
         assert.strictEqual(window.hit, 'FIRED', 'Hit was fired');
         assert.strictEqual(window.sessionStorage.getItem(cName), 'rejected', 'sessionStorage item has been set');
         clearStorageItem(cName);
+
+        cName = '__test-item_allowed';
+        cValue = 'allowed';
+        runScriptlet(name, [cName, cValue]);
+        assert.strictEqual(window.hit, 'FIRED', 'Hit was fired');
+        assert.strictEqual(window.sessionStorage.getItem(cName), 'allowed', 'sessionStorage item has been set');
+        clearStorageItem(cName);
+
+        cName = '__test-item_denied';
+        cValue = 'denied';
+        runScriptlet(name, [cName, cValue]);
+        assert.strictEqual(window.hit, 'FIRED', 'Hit was fired');
+        assert.strictEqual(window.sessionStorage.getItem(cName), 'denied', 'sessionStorage item has been set');
+        clearStorageItem(cName);
     });
 
     test('Set sessionStorage key with invalid value', (assert) => {